compileSummary(MSToolkit)	R Documentation

Compile all Micro or Macro evaluation data

Description

This function reads all of the micro evaluation or macro evaluation data and amalgamates it into a single data frame which has an additional column indicating which replicate each row belongs to. This data frame is then written to "m[ia]croSummary.csv".

Usage


compileSummary(dataType = c("MicroEvaluation", "MacroEvaluation"), replicates = NULL, tryPerl = TRUE, workingPath = getwd())

Arguments

dataType 	 (Optional) A string that indicates which of the two data types should be read. Must be either "MicroEvaluation" or "MacroEvaluation". Partial matching is used here. See match.arg. Default is "MicroEvaluation"

replicates 	 (Optional) Which replicates should be compiled. By default, the function will compile all available replicate data at the time

tryPerl 	 (Optional) Should the alternative "Perl" script be tried (faster). The perl script will ony be used if perl in available on the system, most likely if it is a unix system. Default is TRUE

workingPath 	 (Optional) The working directory to be used as the root for the "Micro" and "Macro" evaluation directories. Also the directory in which the compiled summary files will be created. The current working directory is used by default

Value

CompileSummary does not return a value.

Author(s)

Mango Solutions <mstoolkit@mango-solutions.com>

Examples


  ## Not run: 
    # a directory MicroEvaluation need to be there and be full 
    # of files matching the pattern micro[0-9]{4}.csv
    compileSummary(dataType = "Micro")
    
    # a directory MacroEvaluation need to be there and be full 
    # of files matching the pattern macro[0-9]{4}.csv
    compileSummary(dataType = "Macro")
    
    # the perl script
    file.show( system.file( "exec", "compile", package = "MSToolkit" ))

  ## End(Not run)

[Package MSToolkit version 1.5-1 Index]