Create replicate data, micro evaluation, and macro evaluation directories

This function creates directories for holding data for the current scenario. These directories will hold the replicate data, the micro evaluation data and the macro evaluation data.

Usage

createDirectories(
  dirNames = c("ReplicateData", "MicroEvaluation", "MacroEvaluation"),
  workingPath = getwd(),
  warn = FALSE,
  method = getEctdDataMethod()
)

Arguments

dirNames

(Optional) A vector of directories to create under the "workingPath" directory. By default, directories "ReplicateData", "MicroEvaluation" and "MacroEvaluation" are created

workingPath

(Optional) The working directory in which to create directories. The current working directory is used by default

warn

(Optional) A logical value. Should warnings from the directory creation be shown? FALSE by default

method

Data storage method (given by getEctdDataMethod by default). MSToolkit can work with externalised data (as .csv or .RData files) or internal data format, where the replicate data is stored as a list of data frames in an internal environment (.ectdEnv$DataStore)

Value

A logical vector the same length as "dirNames", indicating whether or not the corresponding directories were successfully created.

Details

Attempts to create directories named in character vector "dirNames" within the "workingPath" directory. No action is taken if method is "Internal"

Author

Francisco Gochez

Examples

  if (FALSE) {
    # Create 2 of the 3 directories
    createDirectories(dirNames = c("ReplicateData", "MicroEvaluation"))
  }