createExternalCovariates(MSToolkit)	R Documentation

Create covariates by sampling from an external file

Description

Create covariates by sampling from an external file.

Usage


createExternalCovariates(subjects, names, file, sameRow = TRUE, 
  subset = NULL, refCol = NULL, dataId = idCol, idCol = "SUBJ", 
  percent = 20, seed = .deriveFromMasterSeed(), includeIDCol = TRUE, 
  refColSuffix = "refCol", workingPath = getwd())

Details

The sampling is always done with replacement.

The refCol setting is typically used with the parameter component, see createParameters or createExternalParameters to maintain consistency between imported covariates and imported parameters.

Value

A data frame containing the imported variables, and possibly a reference variable.

Author(s)

Mango Solutions <mstoolkit@mango-solutions.com>

See Also

createContinuousCovariates, createDiscreteCovariates, and createCovariates

Examples

## Not run: 
  
  # an example file from the unit tests of the MSToolkit package
  wPath <- system.file( "Runit", "data", "createCovariates", package = "MSToolkit" ) 
  
  # sample 20 subjects from the example file 
  dat <- createExternalCovariates( 20, names = "X1", 
    subset = c(".7 < X1 < .8", "-1 <= X2 <= 1"),  
    file = "testCovariates.csv", workingPath = wPath )
  print( dat )
    
  # maintaining the origin of each row  
  dat <- createExternalCovariates( 20, names = "X1, X2", 
    subset = c(".7 < X1 < .8", "-1 <= X2 <= 1"),  
    file = "testCovariates.csv", workingPath = wPath, refCol = "ID" )
  print( dat )

## End(Not run)

[Package MSToolkit version 1.5-1 Index]