Creates parameters for subjects in each replicate

Creates simulated fixed and between subject parameters for subjects in each replicate

Usage

createParameters(
  subjects,
  genNames,
  genFixedMean,
  genFixedCov = 0,
  genRange,
  genBetweenNames,
  genBetweenMean,
  genBetweenCov,
  genErrStruc = "None",
  genMaxDraws = 10,
  genParRangeTolerance = 0.5,
  extFile,
  extNames,
  extBetween,
  extBetweenNums,
  extSubset,
  extRange,
  extErrStruc = "None",
  extRefCol,
  extRefColName,
  extRefColSuffix,
  extIndEffects,
  extDataId = idCol,
  workingPath = getwd(),
  suffix = ".Between",
  idCol = getEctdColName("Subject"),
  seed = .deriveFromMasterSeed(),
  flagName = getEctdColName("ParOmit")
)

Arguments

subjects

Subjects for which to create parameters

genNames

Names of fixed effects to generate

genFixedMean

Means for generating fixed parameters

genFixedCov

Covariance Matrix for generating fixed parameter. 0 by default.

genRange

Range of Acceptable values for derived parameters

genBetweenNames

Between subjects effects to generate

genBetweenMean

Means for generated between subject effects

genBetweenCov

Between subject effects covariance.

genErrStruc

The function to map effects: Additive, Proportional, Log-Normal or None. The default is "None".

genMaxDraws

The maximum number of draws. The default value is 10.

genParRangeTolerance

The proportion of subjects with "in range" parameters with. The default value is 0.5.

extFile

The external file name for data to import.

extNames

Names of parameters to import (refers to variables in the input data)

extBetween

Between subject effects variables in the data

extBetweenNums

The integer mapping between random and fixed effects.

extSubset

The subset to be applied to data before sampling.

extRange

The range of Acceptable values for derived parameters.

extErrStruc

Function to map between subject effects: Additive, Proportional, Log-Normal or None. The default is "None".

extRefCol

The column of reference data.

extRefColName

The column name in data for referenced sampling.

extRefColSuffix

Suffix to add to reference variable in the data.

extIndEffects

Individual effects flag

extDataId

The external subject variable name. By default, this is taken to be the same as the "idCol" input.

workingPath

Working path from which to import covariate file. By default, the working path will be used

suffix

Suffix for retained between subject effects. The default is ".Between".

idCol

Subject variable name. The default is getEctdColName("Subject").

seed

Random seed. The default is .deriveFromMasterSeed().

flagName

Name for omit flag. The default is getEctdColName("ParOmit").