createResponse(MSToolkit)	R Documentation

High-level function for the response component.

Description

Create a response from a dataset and an equation.

Usage


createResponse(data, equation, name = "RESP", invLink, distribution = "normal", 
  covariance, errStruc = "additive", range, digits = 3, 
  seed, flagName = "RESPOMIT")

Arguments

data 	 (Required) Data structure to which to add response.

equation 	 (Required) R function or character string for creating response. See createResponseVariable.

name 	 (Optional) Response variable name ("RESP" by default)

invLink 	 (Optional) Inverse link function for the predictor. This is an R function. No inverse link function by default

distribution 	 (Optional) Outcome variable distribution, one of "n" for normal, "l" for lognormal, "b" for binomial or "p" for poisson. Default is "Normal"

covariance 	 (Optional) Residual Error (CO)VARIANCE, passed to the addResidualError function. Missing by default, resulting in no residual error generated for the reponse

errStruc 	 (Optional) Function describing how to apply residual error: "Additive" (default) or "Proportional"

range 	 (Optional) Range of acceptable values for created response. See parseRangeCode. Missing by default, resulting in no "range" limitation on the response

digits 	 (Optional) Number of digits to which to round the response (3 by default)

seed 	 (Optional) Random number seed. By default, this is derived from the current random seed

flagName 	 (Optional) "Response Omit" Flag name ("RESPOMIT" by default)

Value

A data set containing two columns.

RESP 	 The actual Response

RESPOMIT 	 The response OMIT flag. It is named from the value of the flagName argument.

Author(s)

Mango Solutions <mstoolkit@mango-solutions.com>

See Also

The lower-level function of the response component: createResponseVariable and addResidualError are called by createResponse.

Examples


  # example data
  myData <- expand.grid( X = 1:2, Y = 1:2, Z = 1:2 )
  
  # function example
  out1 <- createResponse(data = myData, equation   = function(data) with(data, X+Y+Z), 
    covariance = 1, range = "RESP < 3", seed = 96)
  
  # same example using a character representation
  out2 <- createResponse(data = myData, equation   = "X+Y+Z", 
    covariance = 1, range = "RESP < 3", seed = 96)
  stopifnot( identical( out1, out2) )  

[Package MSToolkit version 1.5-1 Index]