readData(MSToolkit)	R Documentation

Read scenario data

Description

Reads a single element of replicate, macro evaluation or micro evaluation data from the current scenario

Usage


readData(dataNumber, dataType = c("ReplicateData", "MicroEvaluation", "MacroEvaluation"), variables = NULL, workingPath = getwd())

Arguments

dataNumber 	 (Required) The number of the data entry you wish to read. This number must be between 0 and 9999

dataType 	 (Optional) The type of data you wish to read - either "ReplicateData", "MacroEvaluation" or "MicroEvaluation". By default, "ReplicateData" is used

variables 	 (Optional) The variables that must be in the data to continue. No variables are provided by default (so no variable check is performed)

workingPath 	 (Optional) The working path of the current scenario. By default, the current working directory is used

Value

readData will return a data frame containing the desired element of data if it exists.

Author(s)

Mango Solutions <mstoolkit@mango-solutions.com>

See Also

writeData

Examples


        ## Not run: 
   readData(dataNumber = 10, dataType = "Macro")
         readData(dataNumber = 1, dataType = "Replicate")
         readData(dataNumber = 800, dataType = "Micro")
 
## End(Not run)

[Package MSToolkit version 1.5-1 Index]