removeDirectories(MSToolkit)	R Documentation

Remove directories

Description

This function removes directories containing data for the current scenario.

Usage


removeDirectories(dirNames = c("ReplicateData", "MicroEvaluation", "MacroEvaluation"), workingPath = getwd())

Arguments

dirNames 	 (Optional) A vector of directories to remove from the "workingPath" directory. Default is to remove "ReplicateData", "MicroEvaluation" and "MacroEvaluation" directories

workingPath 	 (Optional) The working directory from which to remove the directories. By default, directories are removed from the current working directory

Details

dirNames must be a character vector. The function will attempt to remove each of the directories identified in the dirNames vector from within the "workingPath" directory

Value

A logical vector the same length as the dirNames, indicating whether or not the corresponding directories were successfully removed

Author(s)

Mango Solutions <mstoolkit@mango-solutions.com>

Examples


  ## Not run: 
    createDirectories(dirNames = c("ReplicateData", "MicroEvaluation"))
    removeDirectories(dirNames = c("ReplicateData", "MicroEvaluation"))
  
## End(Not run)

[Package MSToolkit version 1.5-1 Index]