validNames(MSToolkit)	R Documentation

Check if a character string is a valid R variable name

Description

Checks if a character string is a valid R variable name using regular expressions.

Usage


  validNames(...)

Arguments

... 	 (Required) A collection of character vectors to be checked

Details

A valid R name starts with a letter or a dot followed by a non-numerical character and contains only letters, numbers and dots.

Value

None. The function is simply generates an error if an invalid name is passed to it.

Author(s)

Mango Solutions <mstoolkit@mango-solutions.com>

Examples


  # correct name
  validNames("DOSE")
  ## Not run: 
    # wrong name
    validNames(".334fsedqw")
  
## End(Not run)

[Package MSToolkit version 1.5-1 Index]